Removed broken button
On the theme, there is a button in the Lyrics title with a tooltip that says "Show biography" that causes the size of the lyrics panel to decrease. However, the button doesnt display correctly and only becomes visible when the lyrics panel is at a specific width, although it can be clicked on if it is not visible.

It could serve the purpose of showing the biography if the lyrics panel is expanded to take up the whole space, but its impossible to do that as the width buttons loop back around. As such, it doesn't serve any purpose at the moment and I assume it is either leftover code or its functionality was broken at some point, and I simply commented it out.