Added Look-up button icon fallback
The current version of Biography requires FontAwesome to be installed on the system for some icons to display correctly. If it is not installed, what should be an arrow for the Look-up button will display the Unicode symbol for an unknown character.
This instead will use the Unicode symbol '˅', which will display on all systems, if FontAwesome is not installed.